Reorganize Reconciler code ordering
The vpa.Reconciler function read a bit funny to me: a bunch of private
methods were listed before the exported methods. When someone is going
to read that file, they most likely want to see the public API of the
struct before they see the private functions that implement it. This
commit reorders things so that the public API is listed first in the
file.