Add DryRun field to the Reconciler
The vpa.Reconciler currently uses a passthrough dry-run variable.
However the dry run behavior is part of the state of the reconciler, and
it shouldn't change within the lifecycle of the reconciler. Decoupling
it from the reconciler gives the appearance that the dry run behavior
should somehow be separate, and results in a variable that has to be
passed around to private package functions.