Refactor Reconciler.ReconcileNamespace
The main exported API on of the vpa package is the
Reconciler.ReconcilerNamespace method. However, this method is a bit
hard to read because it mixes levels of abstraction quite frequently.
This leads to comments being used to convey semantic information. This
commit refactors the main reconciliation loop into a series of
unexported methods, allowing the reader of the code to get a better
sense of why certain operations are taking place.