Hkatz/refactor reconciler (#128)
* Revert "Make units consistent for memory fixes issue #78 (#103)"

This reverts commit e55026dd67d00b3c9f68946216ac9df09c6d4a9a.

* Add FormatResourceList and test cases

* Use v1 resource memory constants instead of hardcoded strings

* Reorganize Reconciler code ordering

The vpa.Reconciler function read a bit funny to me: a bunch of private
methods were listed before the exported methods. When someone is going
to read that file, they most likely want to see the public API of the
struct before they see the private functions that implement it. This
commit reorders things so that the public API is listed first in the
file.

* Add DryRun field to the Reconciler

The vpa.Reconciler currently uses a passthrough dry-run variable.
However the dry run behavior is part of the state of the reconciler, and
it shouldn't change within the lifecycle of the reconciler. Decoupling
it from the reconciler gives the appearance that the dry run behavior
should somehow be separate, and results in a variable that has to be
passed around to private package functions.

* Make list/create/deleteVPA Reconciler methods

These functions all take data clumps of internal state of the
Reconciler. While decoupling them might be useful if they were exported
package functions, the fact that they're unexported means that the
Reconciler is their only caller. Given those two facts, it makes more
sense to couple the these functions with the Reconciler.

* Rename checkNamespaceLabel to namespaceIsManaged

The name checkNamespaceLabel describes what is done mechanically, not
why. namespaceIsManaged should hopefully give a clearer sense of what
the boolean return value means for this method.

* Refactor Reconciler.ReconcileNamespace

The main exported API on of the vpa package is the
Reconciler.ReconcilerNamespace method. However, this method is a bit
hard to read because it mixes levels of abstraction quite frequently.
This leads to comments being used to convey semantic information. This
commit refactors the main reconciliation loop into a series of
unexported methods, allowing the reader of the code to get a better
sense of why certain operations are taking place.

Co-authored-by: John Turner <jturner@squarespace.com>
Co-authored-by: Andrew Suderman <andrew@sudermanjr.com>