Hkatz/add vpa update mode (#137)
* Revert "Make units consistent for memory fixes issue #78 (#103)"

This reverts commit e55026dd67d00b3c9f68946216ac9df09c6d4a9a.

* Add FormatResourceList and test cases

* Use v1 resource memory constants instead of hardcoded strings

* Add a VPAUpdateMode to the vpa.Reconciler

The VPAUpdateMode field will allow us to use Goldilocks to create VPAs
that operate outside of "off" mode.

* Add ability to specify vpa update mode

This commit adds a parameter to createVPA that allows the user to
specify the update mode for the VPA. The mechanism for specifying this
is the label goldilocks.fairwinds.com/vpaUpdateMode.

* Extract variables instead of hardcoded values

* VPA update mode label case insensitive

Kubernetes labels are based on DNS names, which are case insensitive.
However, the current label used to specify the update mode of created
VPAs, vpaUpdateMode, is case sensitive. This has caused a bug that
results in all VPAs being created in "Off" mode. This commit fixes that
bug by using a case-insensitive label to specify the vpa-update-mode.

* Summary fix for multiple containers in a pod (#123)

* go.mod updated for some reason

* Add support for all VPA UpdateModes

* Use ParseBool for deployment enabled label value

* Use ParseBool for namespace enabled label value

* Use lowercase for switch statement

Co-authored-by: John Turner <jturner@squarespace.com>
Co-authored-by: David Noyes <david.noyes@skybettingandgaming.com>